Implement logic for simple keyword tagging

+++ core/views.py
@@ -1,9 +1,7 @@
-import asyncio
 import json
 import threading
 from io import BytesIO
 
-from asgiref.sync import sync_to_async, async_to_sync
 from django.shortcuts import redirect, render
 from django.http import JsonResponse
 from django.views.generic import DetailView, ListView
@@ -57,131 +55,149 @@ class WorkflowDetailView(DetailView):
         context = super().get_context_data(**kwargs)
         workflow = Workflow.objects.filter(id=self.kwargs['pk']).first()
         if workflow:
-            context["tags"] = json.loads(workflow.result).get("generated_tags", [])
+            context["result"] = json.loads(workflow.result)
             context["pending"] = PENDING
             context["completed"] = COMPLETED
             context["upload"] = UPLOAD
+            context["process"] = PROCESS
             context["can_abort_status"] = CAN_ABORT_STATUS
         return context
 
 
-@sync_to_async
-def async_get_uid(request):
-    return request.session.get("uid")
-
-
-async def _start_workflow_task(request, wid, file_obj):
-    workflow = await Workflow.objects.filter(id=wid).afirst()
+def start_workflow_task(request, wid, file_obj):
+    workflow = Workflow.objects.filter(id=wid).first()
     if not workflow:
         return
 
-    uid = await async_get_uid(request)
+    uid = request.session.get('uid')
     title = request.POST.get("title")
 
     work_type = int(req_type) if (req_type := request.POST.get("type")).isnumeric() else -1
     if work_type == UPLOAD:
-        paper = await Paper.objects.filter(title=title, owner_id=uid).afirst()
+        paper = Paper.objects.filter(title=title, owner_id=uid).first()
         if not paper:
             return
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         if not workflow or workflow.status == ABORTED:
             return
         workflow.stage = S_UPLOADING
-        await workflow.asave()
+        workflow.save()
 
         cdn_client = SakanaCDNClient()
-        file_path = await cdn_client.store_paper(file_obj)
+        replace = not not request.POST.get("replace")
+        if not replace:
+            file_path = cdn_client.store_paper(file_obj)
+        else:
+            filepath = paper.file_path
+            file_path = cdn_client.replace_paper(filepath, file_obj)
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         # if workflow is aborted or deleted (not allowed for users)
         if not workflow or workflow.status == ABORTED:
-            _ = await cdn_client.delete_paper(file_path)  # delete new paper
-            paper = await Paper.objects.filter(title=title, owner_id=uid).afirst()
+            _ = cdn_client.delete_paper(file_path)  # delete new paper
+            paper = Paper.objects.filter(title=title, owner_id=uid).first()
             if not paper:  # if paper is deleted
                 return
             if not paper.file_path:  # delete paper entry
-                await paper.adelete()
+                paper.delete()
             return
         # else
-        paper = await Paper.objects.filter(title=title, owner_id=uid).afirst()
+        paper = Paper.objects.filter(title=title, owner_id=uid).first()
         if not paper:  # if paper is deleted or altered during normal workflow execution
             workflow.result = json.dumps({**json.loads(workflow.result), **{"error": "paper is not available"}})
             workflow.status = FAILED
-            await workflow.asave()
+            workflow.save()
         old_path = paper.file_path
         if old_path:
             # delete old paper because user must have selected "replace" in this position
-            _ = await cdn_client.delete_paper(old_path)
+            _ = cdn_client.delete_paper(old_path)
         paper.file_path = file_path
-        await paper.asave()
+        paper.save()
 
         workflow.stage = S_END
         workflow.status = COMPLETED
-        await workflow.asave()
+        workflow.save()
 
     elif work_type == PROCESS:
         pid = request.POST.get("pid")
-        paper = await Paper.objects.filter(id=pid, owner_id=uid).afirst()
+        paper = Paper.objects.filter(id=pid, owner_id=uid).first()
         if not paper:
             return
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         if not workflow or workflow.status == ABORTED:
             return
         workflow.stage = S_PROCESSING_0
-        await workflow.asave()
+        workflow.save()
 
         # Task 0: retrieve paper
         file_path = paper.file_path
-        cdn_client = PseudoCDNClient()
-        file_obj = await cdn_client.request_for_paper(file_path)
+        cdn_client = SakanaCDNClient()
+        file_obj = cdn_client.request_for_paper(file_path)
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         if not workflow or workflow.status == ABORTED:
             return
         workflow.stage = S_PROCESSING_1
-        await workflow.asave()
+        workflow.save()
 
         # Task 1: read and process paper
         tags = request.POST.getlist("tag-names")
+        tags_dict = {}  # key: tag name, value: definition
+        for t in tags:
+            tag = Tag.objects.filter(name=t).first()
+            tags_dict[t] = tag.definition
         llm_client = SimpleKeywordClient()
-        matching_tags = await llm_client.match_paper_on_tags(file_obj, tags)
+        matching_tags = llm_client.match_paper_on_tags(file_obj, tags)
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         if not workflow or workflow.status == ABORTED:
             return
         workflow.stage = S_PROCESSING_2
-        await workflow.asave()
+        workflow.save()
 
-        # Task 2: tag paper
-        paper = await Paper.objects.filter(id=pid, owner_id=uid).afirst()
+        # Task 2: tag paper (add matching tags, remove non-matching)
+        paper = Paper.objects.filter(id=pid, owner_id=uid).first()
         if not paper:
             return
+        original_tags = [t.name for t in paper.tags.all()]
+        added_tags = []
+        removed_tags = []
+        kept_tags = []
         for t in matching_tags:
-            tag = await Tag.objects.filter(name=t).afirst()
+            if t in original_tags:
+                kept_tags.append(t)
+            else:
+                added_tags.append(t)
+            tag = Tag.objects.filter(name=t).first()
             if tag:
-                await paper.tags.aadd(tag)
+                paper.tags.add(tag)
+        for t in tags:
+            if t in original_tags and t not in matching_tags:
+                tag = Tag.objects.filter(name=t).first()
+                if tag:
+                    paper.tags.remove(tag)
+                removed_tags.append(t)
 
         # in case of caching, re-query
-        workflow = await Workflow.objects.filter(id=wid).afirst()
+        workflow = Workflow.objects.filter(id=wid).first()
         if not workflow or workflow.status == ABORTED:
             return
         workflow.stage = S_END
         workflow.status = COMPLETED
-        await workflow.asave()
+        workflow.result = json.dumps({**json.loads(workflow.result), **{"added_tags": added_tags,
+                                                                        "removed_tags": removed_tags,
+                                                                        "kept_tags": kept_tags}})
+        workflow.save()
 
     else:  # illegal work type parameter
         return
-
-
-def start_workflow_task(*args, **kwargs):  # we need to pass a sync function to thread
-    asyncio.run(_start_workflow_task(*args, **kwargs))
 
 
 def handle_create_workflow(request):
@@ -427,18 +443,19 @@ def restore_workflow(request):
     return JsonResponse({"status": 0})
 
 
-@sync_to_async
 @login_required()
-@async_to_sync
-async def delete_paper(request):
-    uid = await async_get_uid(request)
+def delete_paper(request):
+    uid = request.session.get("uid")
     pid = request.POST.get("pid")
-    paper = await Paper.objects.filter(id=pid, owner_id=uid).afirst()
+    paper = Paper.objects.filter(id=pid, owner_id=uid).first()
     if not paper:
         err_msg = "Paper does not exist!"
         return JsonResponse({"status": 1, "err_msg": err_msg})
+    if not paper.file_path:  # could be deleting a paper before the upload finishes
+        err_msg = "Invalid request, please contact the administrator!"
+        return JsonResponse({"status": 1, "err_msg": err_msg})
 
-    cdn_client = PseudoCDNClient()
-    _ = await cdn_client.delete_paper(paper.file_path)
-    await paper.adelete()
+    cdn_client = SakanaCDNClient()
+    _ = cdn_client.delete_paper(paper.file_path)
+    paper.delete()
     return JsonResponse({"status": 0})
